refactor(auth): move auth command logic into folder
Move the auth command manager, login flow, verification helpers, and tests into internal/cli/auth so the subcommand folder owns its command-specific behavior. Keep only the shared platform base-URL normalization helper in internal/cli for reuse by platform_client.
